+++ changelog.md
@@ -1,3 +1,4 @@
+- Add Sandstone trims (#35)
 - Fix benches being present in sheets tag
 - Fix incorrect light blue and light gray sheet and cushion recipes (#66)
 - When placing a cushion/sheet on furniture, you can now replace the existing cushion/sheet without having to take off
@@ -5,3 +6,5 @@
 - Add hammer clarification tooltip to counters
 - Small optimization to fancy painting rendering
 - Fix dark oak palette to match vanilla
+- Add bamboo and spruce pillar trim variants
+- Fix corner trim having one side less thick (#68)